Create min_changes_to_y.py
Given a square matrix n × n (n is guaranteed to be odd) that contains only 0s, 1s, and 2s. In one operation, you are allowed to change the number in any cell of the matrix to a different number (0, 1, or 2). Your task is to compute the minimum number of cells that need to change for the letter Y to be written on the matrix.

The letter Y is written on the matrix if and only if:

all numbers on the diagonals starting from the upper-left and upper-right corners down to the center of the matrix as well as the numbers stretching down vertically from the center of the matrix are equal;
all other numbers that are not part of that Y should be equal and be different from the numbers that make up the Y.
Note: for a square matrix of size n × n, there are exactly 6 possible ways of writing Y on matrix. The cells that make up the Y letter and the cells that make up its background can be equal to 0 and 1, 0 and 2, 1 and 0, 1 and 2, 2 and 0, and 2 and 1, respectively.


Example

For

matrix = [
  [1, 0, 2],
  [1, 2, 0],
  [0, 2, 0]
]
the output should be solution(matrix) = 2.

Explanation:

The best is to change the 1 in the 0th row to 2 and the 1 in the 1st row to 0.
2s make the Y letter and 0s make its background.
The final matrix is shown below.
matrix = [
  [2, 0, 2],
  [0, 2, 0],
  [0, 2, 0]
]