Create Calculate_diff_nums_of_uppercase_and_lowercase.py
In a typing practice application, you need to validate the consistency of user input. Given a string typedText that includes uppercase and lowercase English letters, you need to compute the difference between the number of uppercase and lowercase letters in this string. Return the difference as an integer.

Note: You are not expected to provide the most optimal solution, but a solution with time complexity not worse than O(typedText.length2) will fit within the execution time limit.

Example

For typedText = "CodeSignal", the output should be solution(typedText) = -6.

Explanation:
There are 2 uppercase letters ('C', 'S') and 8 lowercase letters ('o', 'd', 'e', 'i', 'g', 'n', 'a', 'l').
The difference is 2 - 8 = -6.

For typedText = "a", the output should be solution(typedText) = -1.

Explanation:
There are no uppercase letters and 1 lowercase letter ('a').
The difference is 0 - 1 = -1.

For typedText = "AbCdEf", the output should be solution(typedText) = 0.

Explanation:
There are 3 uppercase letters ('A', 'C', 'E') and 3 lowercase letters ('b', 'd', 'f').
The difference is 3 - 3 = 0.